backend/requirements.txt
# requirements.txt
Flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
user-agents==2.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.11
psycopg2